在服务器完成了流程，1.0

+++ application/Aorta_nii2dcm.py
@@ -29,11 +29,17 @@ def copy_ct(extract_path=r"F:\ZS_Aorta\output", target_path=r"D:\Aorta", r=None)
     for i, folder_path in enumerate(os.listdir(extract_path)):
         if r[0] <= i <= r[1]:
             extract_fpath = os.path.join(extract_path, folder_path + "\\ct.nii")
+
+            target_folder = os.path.join(target_path, folder_path)
+            if not os.path.exists(target_folder):
+                os.makedirs(target_folder)
+
             target_fpath = os.path.join(target_path, folder_path + "\\ct.nii")
-            try:
-                shutil.copyfile(extract_fpath, target_fpath)
-            except FileNotFoundError:
-                print(extract_fpath, " Not Found")
+            if not os.path.isfile(target_fpath):
+                try:
+                    shutil.copyfile(extract_fpath, target_fpath)
+                except FileNotFoundError:
+                    print(extract_fpath, " Not Found")
 
 
 def generate_one_series(read_dir, save_dir, patient_name):
@@ -41,7 +47,7 @@ def generate_one_series(read_dir, save_dir, patient_name):
     g.Execute(out_dir=save_dir, PatientName=patient_name)
 
 
-def generate_all_series(base_folder):
+def generate_all_series(base_folder=r"D:\Aorta"):
     for folder in os.listdir(base_folder):
         fpath = base_folder + "\\" + folder + "\\ct.nii"
         dcm_folder = base_folder + "\\" + folder + "\\dcm"
@@ -49,7 +55,7 @@ def generate_all_series(base_folder):
         g.Execute(out_dir=dcm_folder, PatientName=folder)
 
 
-def copy_dcm_to_auto(base_folder, target_path=r"D:\DeepViewer\data\pacs"):
+def copy_dcm_to_auto(base_folder=r"D:\Aorta", target_path=r"D:\DeepViewer\data\pacs"):
     """
     将dcm文件从总文件夹中复制到自动处理文件夹
     :param base_folder:
@@ -60,8 +66,11 @@ def copy_dcm_to_auto(base_folder, target_path=r"D:\DeepViewer\data\pacs"):
         dcm_folder = base_folder + "\\" + folder + "\\dcm"
         for fname in os.listdir(dcm_folder):
             fpath = os.path.join(dcm_folder, fname)
-            shutil.copyfile(fpath, target_path)
+            shutil.copyfile(fpath, os.path.join(target_path, fname))
 
 
 if __name__ == "__main__":
-    copy_ct(r=(0, 5))
+    # copy_ct(r=(0, 5))
+    # generate_all_series()
+    copy_dcm_to_auto()
+
